updated test script

+++ src/tst.sh
@@ -1,14 +1,19 @@
 #! /bin/sh
 
-cd ../tests
+if test -d ../tests;
+then
+    cd ../tests
 
-\rm -f *.diff
-\rm -f *.tmp
-\rm -f *.out*
+    \rm -f *.diff
+    \rm -f *.tmp
+    \rm -f *.out*
 
-for file in test2text testtxt testalto testcorrect \
-	    testhocr testidf testpage testlangcat teststats \
-	    testcollect testpm \
-	    testclean testwordtranslate
-do bash ./testone.sh $file
-done
+    for file in test2text testtxt testalto testcorrect \
+		testhocr testidf testpage testlangcat teststats \
+		testcollect testpm \
+		testclean testwordtranslate
+    do bash ./testone.sh $file
+    done
+else
+    exit 1
+fi
